fix: address PR review findings - fix teamDescription validation to count characters, not blocks - remove duplicate nested .team-description wrapper and empty class attr - extract hasSanityImage generic guard to client.ts - DRY portableTextPlainLength into shared studio utility - replace unsupported max-width: stretch with 100% - add _id path test coverage for team logo guard
