+++ src/lib/sanity/client.ts
@@ -17,3 +17,13 @@ const builder = createImageUrlBuilder(sanityClient);
 export function urlFor(source: SanityImage) {
   return builder.image(source);
 }
+
+/** True when a Sanity image field has a resolvable asset (optional fields may be unset or empty). */
+export function hasSanityImage(
+  image: SanityImage | null | undefined,
+): image is SanityImage {
+  if (!image || typeof image !== "object") return false;
+  const asset = (image as { asset?: { _ref?: string; _id?: string } }).asset;
+  if (!asset || typeof asset !== "object") return false;
+  return Boolean(asset._ref ?? asset._id);
+}
