+++ studio-hack-michigan-/schemaTypes/hackathonProject.ts
@@ -1,26 +1,8 @@
 import {defineField, defineType} from 'sanity'
+import {portableTextPlainLength} from './portableTextUtils'
 
 /** Plain-text character cap for portable text `description` — keep in sync with `src/lib/portableText.ts`. */
 const DESCRIPTION_MAX_CHARS = 5000
-
-function portableTextPlainLength(blocks: unknown): number {
-  if (!Array.isArray(blocks)) return 0
-  let len = 0
-  for (const block of blocks as {_type?: string; children?: unknown[]}[]) {
-    if (block?._type !== 'block' || !Array.isArray(block.children)) continue
-    for (const child of block.children) {
-      if (
-        child &&
-        typeof child === 'object' &&
-        'text' in child &&
-        typeof (child as {text: unknown}).text === 'string'
-      ) {
-        len += (child as {text: string}).text.length
-      }
-    }
-  }
-  return len
-}
 
 export const hackathonProjectType = defineType({
   name: 'hackathonProject',
